finishes placement and routes wires
This change places the raspberry pi header in the
correct location. It also routes wires to all
relevant locations. For some reason the raspberry
pi header we're using wants to connect all of the
power and ground pins together, but that shouldn't
be necessary, so I've left them unconnected and
they show as air wires.